Changed work structure and objeck list for TDualTh class
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@573 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/m_unsp_gen.cpp
@@ -36,7 +36,7 @@ short Prime[]=
 short TUnSpace::near_prime_number( short num )
 {
   int i;
-  for( i=0; i<sizeof(Prime)/sizeof(short)-1; i++)
+  for( i=0; i<(int)(sizeof(Prime)/sizeof(short))-1; i++)
    if( Prime[i] >= num )
            break;
